Improvements based on comments.
